feat: Upd Timer(double); Add ./scripts/extract_twoColumn.py

+++ include/basis.hpp
@@ -68,7 +68,7 @@ class Timer {
 #endif
   }
 
-  // 回傳奈秒
+  // 回傳奈秒(有無小數)
   __attribute__((always_inline)) inline unsigned long long elapsed_ns() const {
 #if TIMER_METHOD == TIMER_STEADY_CLOCK
     return std::chrono::duration_cast<std::chrono::nanoseconds>(Clock::now() -
@@ -76,8 +76,10 @@ class Timer {
         .count();
 #elif TIMER_METHOD == TIMER_RDTSCP
     ensure_cpu_freq_hz();
-    return static_cast<unsigned long long>((perf_counter() - m_beg) *
-                                           (1e9 / cpu_freq_hz.load()));
+    // return static_cast<unsigned long long>((perf_counter() - m_beg) * (1e9 /
+    // cpu_freq_hz.load()));
+    return static_cast<double>(perf_counter() - m_beg) *
+           (1e9 / cpu_freq_hz.load());
 #endif
   }
